Add OpenMP support.
Note that OpenBlas (or intel MKL) already use OpenMP to optimise some subroutines. It is therefore not possible to optimise those parts without creating a two stages thread initializer. Depending on the compiler, it may throw an error:
"Program is Terminated. Because you tried to allocate too many memory regions"